+++ src/Game.Abstractions/GameComponent.cs
@@ -14,7 +14,7 @@
         {
         }
 
-        public virtual void Update(double delta)
+        public virtual void Update(float delta)
         {
         }
     }
