+++ src/Game.Abstractions/IGameComponent.cs
@@ -5,6 +5,6 @@
         void Load();
         void Quit();
         void Draw();
-        void Update(double delta);
+        void Update(float delta);
     }
 }
